BookAPI files added as a SpringBoot app
