+++ simon8491/main/java/pl/coderslab/book_api/BookApiApplication.java
@@ -1,0 +1,12 @@
+package pl.coderslab.book_api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BookApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BookApiApplication.class, args);
+	}
+}
