+++ simon8491/main/java/pl/coderslab/book_api/ServletInitializer.java
@@ -1,0 +1,13 @@
+package pl.coderslab.book_api;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(BookApiApplication.class);
+	}
+
+}
