+++ simon8491/main/java/pl/coderslab/book_api/model/Book.java
@@ -1,0 +1,19 @@
+package pl.coderslab.book_api.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class Book {
+    private Long id;
+    private String isbn;
+    private String title;
+    private String author;
+    private String publisher;
+    private String type;
+}
