+++ simon8491/main/resources/application.properties
@@ -1,0 +1,1 @@
+spring.application.name=book-api
