--- simon8491/src/main/java/pl/coderslab/AppInitializer.java
@@ -1,39 +1,0 @@
-package pl.coderslab;
-
-import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRegistration;
-import org.springframework.stereotype.Component;
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
-import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
-
-//@Component
-//public class AppInitializer implements WebApplicationInitializer {
-//    @Override
-//    public void onStartup(ServletContext servletContext) throws ServletException {
-//        System.out.println(">>> AppInitializer is running");
-//        AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-//        context.register(AppConfig.class);
-//
-//        ServletRegistration.Dynamic dispatcher = servletContext.addServlet("dispatcher", new DispatcherServlet(context));
-//        dispatcher.setLoadOnStartup(1);
-//        dispatcher.addMapping("/");
-//    }
-//}
-
-
-public class AppInitializer extends
-        AbstractAnnotationConfigDispatcherServletInitializer {
-    @Override
-    protected Class<?>[] getRootConfigClasses() { return null; }
-
-    @Override
-    protected Class<?>[] getServletConfigClasses() {
-        return new Class[]{AppConfig.class};  }
-
-    @Override
-    protected String[] getServletMappings() {
-        return new String[]{"/"};   }
-}
