--- simon8491/src/main/resources/META-INF/services/jakarta.servlet.ServletContainerInitializer
@@ -1,1 +1,0 @@
-org.springframework.web.SpringServletContainerInitializer
