+++ simon8491/test/java/pl/coderslab/book_api/BookApiApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.coderslab.book_api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BookApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
